now does not show cmd windows anymore; created one method in cmdTools.go to exec them all
